*** empty log message ***
git-svn-id: file:///home/chulwoo/CPS/repo/CVS2SVN/cps/trunk@2984 fb734a29-28db-4bc5-88a3-3cb5ee595545

+++ cps_pp/include/util/Flconv.h
@@ -6,7 +6,8 @@
 #ifndef __FCONVERT__
 #define __FCONVERT__
 
-typedef unsigned long int type32 ;
+//typedef unsigned long int type32 ;
+typedef uint32_t type32 ;
 
 void ti2ieee (type32 *,int);
 //void ieee2ti (type32 *,int);
